calendar_agent_a2a/pyproject.toml
[project]
name = "a2ui-calendar-agent"
version = "0.1.0"
description = "Calendar agent with Google ADK that uses A2UI and is hosted as an A2A server agent."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "a2ui @ file:///Users/shawnkelshaw/Windsurf/Google%20AI%20Studio%20Projects/AdventofAgents/A2UI/a2a_agents/python/a2ui_extension",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
